Adding navigation components for main 3 screens: login, register, adn start
