+++ app/src/main/java/com/example/soulsync/navigation/AppDestinations.kt
@@ -1,0 +1,7 @@
+package com.example.soulsync.navigation
+
+sealed class AppDestinations(val route: String) {
+    object StartScreen : AppDestinations("startScreen")
+    object LoginScreen : AppDestinations("loginScreen")
+    object RegisterScreen : AppDestinations("registerScreen")
+}
