--- app/src/main/java/com/example/soulsync/navigation/AppNavigation.kt
@@ -1,2 +1,0 @@
-package com.example.soulsync.navigation
-
